requirements.txt
aiogram==2.22.1
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.1.1
frozenlist==1.3.1
greenlet==1.1.3
idna==3.3
multidict==6.0.2
pip==22.3.1
python-dotenv==0.20.0
pytz==2022.2.1
requests==2.28.1
setuptools==58.1.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.40
urllib3==1.26.12
yarl==1.8.1
requests-html==0.10.0
alembic==1.9.2
psycopg2==2.9.5